+++ aio_overpass/error.py
@@ -465,10 +465,6 @@ async def __raise_for_html_result(response: aiohttp.ClientResponse, query_kwargs
     if reject_causes:
         raise QueryRejectError(kwargs=query_kwargs, remarks=errors, cause=reject_causes[0])
 
-    # TODO match status code here? f.e. 429 indicates TOO_MANY_QUERIES,
-    #   regardless of whether we actually match an error text
-    #   => it's thinkable that we make too many status requests as well f.e.
-
     raise QueryResponseError(
         response=response,
         body=await response.text(),